requirements.txt
scikit-learn
seaborn
numpy
mlflow
fastapi[standard]
imblearn
sqlalchemy 
sqlmodel
structlog
psycopg2-binary 
python-multipart 
passlib[bcrypt] 
python-jose
pwdlib[argon2]
PyJWT
imbalanced-learn
python-jose[cryptography]
python-json-logger
httpx
pydantic
pytest
azure-storage-blob
